requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
numpy==1.23.5
opencv-python==4.8.0
flask==2.3.3